+++ src/features/canvas/ErdCanvas/index.ts
@@ -1,0 +1,1 @@
+export { default, ERD_CANVAS_ID } from "~/features/canvas/ErdCanvas/ErdCanvas";